Add keyValStore to connection profile in PeerAdmin card
This is for (temporary) compatibility with Playground since it does not (yet)
use the new card-based AdminConnection and BusinessNetworkConnection connect API,
and so needs to have a keyValStore property for the hlfv1 connector to work.
Using the new card-based API, this property will be ignored, so is safe either way.

Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>